dcfdsf

+++ testCardSwiper/CardDetailViewController.h
@@ -22,7 +22,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface StickCollectionViewFlowLayout : UICollectionViewFlowLayout
 @property (nonatomic, assign) CGFloat firstItemTransform;
 @property (nonatomic, assign) NSInteger startIdx;
-@property (nonatomic, assign) CGFloat prePos;
+@property (nonatomic, assign) NSInteger prePos;
 
 @property (nonatomic, assign) CGFloat standardHeight;
 @property (nonatomic, assign) CGFloat featuredHeight;
